+++ src/main/java/me/drex/vanish/mixin/interaction/SimpleCriterionTriggerMixin.java
@@ -1,5 +1,7 @@
 package me.drex.vanish.mixin.interaction;
 
+import com.llamalad7.mixinextras.injector.wrapmethod.WrapMethod;
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import me.drex.vanish.api.VanishAPI;
 import me.drex.vanish.config.ConfigManager;
 import net.minecraft.advancements.critereon.SimpleCriterionTrigger;
@@ -12,13 +14,15 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import java.util.function.Predicate;
 
 @Mixin(SimpleCriterionTrigger.class)
-public abstract class SimpleCriterionTriggerMixin {
-
-    @Inject(method = "trigger", at = @At("HEAD"), cancellable = true)
-    public void preventAdvancementProgress(ServerPlayer player, Predicate<Object> predicate, CallbackInfo ci) {
+public abstract class SimpleCriterionTriggerMixin<T extends SimpleCriterionTrigger.SimpleInstance> {
+    @WrapMethod(method = "trigger")
+    public void preventAdvancementProgress(ServerPlayer player, Predicate<T> predicate, Operation<Void> original) {
         if (ConfigManager.vanish().interaction.advancementProgress && VanishAPI.isVanished(player)) {
-            ci.cancel();
+            return;
         }
+        if (ConfigManager.vanish().interaction.spectatorAdvancementProgress && player.isSpectator()) {
+            return;
+        }
+        original.call(player, predicate);
     }
-
 }