+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `hideGameMode` config option
+- `spectatorAdvancementProgress` config option
 
 ## [1.5.20] - 2025-07-02
 ### Added